Mock Other Components
Create the force-app/test/jest-mocks directory and create mock
components for a lightning namespace override, a separate namespace
called thunder and a component in the default c namespace that is not
included in the force-app/main/default/lwc directory. Modify
jest.config.js with values of mocks in moduleNameMapper property.

Add otherComponents LWC that demonstrates the process of mocking other
components in order to enable Jest testing.