Complete Write Positive Tests unit
Complete the Write Positive Tests unit of the Advanced Apex Testing
module. Added AccountWrapperTests class that tested the
getRoundedAvgPriceOfOpps() method during the unit and isHighPriority()
in the challenge.

In the challege, wrote positive tests for all three equality scenarios
(greater than, equal to, less than).

NOTE: Bad challenge logic as-of January 2, 2024

One of your test wrapper variables MUST be named acctWrapper or the
challenge will fail. This was not specified in the challenge
instructions. Originally had acctWrapper named acmeWrapper and wrapper
named acctWrapper in the loop and the challenge was failing, claiming
that isHighPriority() was never called in the test class.